removing the scripts project from maven...
The scripts project was intended to be part of maven so that the web app could directly invoke the scripts via maven sources dependency, this however failed with openshift and proved to be less useful than direct invocation from the build server.

Groovy has sufficient handling of maven deps that a one-way dependancy can be made with ease should we need to.